+++ src/tree_index/node.rs
@@ -187,7 +187,7 @@ where
                     internal_node.finish_split(barrier);
                     old_root.commit(barrier);
                 }
-                old_root.release(barrier);
+                let _ = old_root.release(barrier);
             };
 
             (key, value)
@@ -239,7 +239,7 @@ where
             match root.compare_exchange(root_ptr, (None, Tag::None), Acquire, Acquire, barrier) {
                 Ok((old_root, _)) => {
                     if let Some(old_root) = old_root {
-                        old_root.release(barrier);
+                        let _ = old_root.release(barrier);
                     }
                     return Ok(true);
                 }